// packages/medusa-react-admin/src/lib/models/AdminPostSalesChannelsSalesChannelReq.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type AdminPostSalesChannelsSalesChannelReq = {
  /**
   * Name of the sales channel.
   */
  name?: string;
  /**
   * Sales Channel description.
   */
  description?: string;
  /**
   * Indication of if the sales channel is active.
   */
  is_disabled?: boolean;
};